requirements.txt
fastmcp==2.13.0.2
psycopg2-binary==2.9.9
openai>=1.54.0
python-dotenv==1.1.1
pyyaml==6.0.1
pydantic>=2.0.0
